+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.4 (12 Jan 2022)
+
+* Added check for alphabet, int and float (+ & -), white space
+* Reverse string
+* Split string into char array
+* Capitalize and unCapitalize string
+
 ## 1.0.3 (7 Jan 2022)
 
 * Changed types in package.json to direct to proper definition as previously I did not point it to index.d.ts